+++ core/server/common/src/main/java/alluxio/master/journal/AbstractCatchupThread.java
@@ -12,17 +12,20 @@
 package alluxio.master.journal;
 
 import alluxio.ProcessUtils;
+import alluxio.thread.AbstractAutopsyThread;
 
-import alluxio.thread.AutopsyThread;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * Abstract helper for tracking and terminating a thread for journal catch-up task.
  */
-public abstract class AbstractCatchupThread extends AutopsyThread {
+public abstract class AbstractCatchupThread extends AbstractAutopsyThread {
   private static final Logger LOG = LoggerFactory.getLogger(AbstractCatchupThread.class);
 
+  /**
+   * Constructor.
+   */
   public AbstractCatchupThread() {
     super();
   }
@@ -57,4 +60,9 @@ public abstract class AbstractCatchupThread extends AutopsyThread {
    * Does the work of catching up.
    */
   protected abstract void runCatchup();
+
+  @Override
+  public void onError(Throwable t) {
+    setError(t);
+  }
 }
